Fixes shotgun choke removal not returning stuns (#8714)
# About the pull request

fixes: #8644
the way the no knockback trait worked is that if the projectile had it
in its bullet_traits list, there would be no knockback. but removing the
choke never removed the trait from the currently chambered projectile,
and the already written code for attachment bullet trait stuff
(traits_to_give) still kept the trait on the projectile for whatever
reason.

also switched the choke to use traits_to_give. i missed that in the
original pr, whoops.

# Explain why it's good for the game

bug fix!! :D


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/user-attachments/assets/181d28c2-662a-41ac-aa6f-2b3c38b1898b


</details>


# Changelog
:cl:
fix: Removing the shotgun choke will return stuns to the currently
chambered projectile.
/:cl: