Adds a "Play as" option to the create xeno/human procs (#8727)
# About the pull request

title

# Explain why it's good for the game

Faster way to quickly inject to a situation 
(im only adding this because it makes testing things easier)
doesnt have a problem with multiple spawns from what ive tested


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
admin: Admins now have a "Play as" option that puts them in the mob they
spawn as.
/:cl:

---------

Co-authored-by: InsaneRed <prodexter31@outlook.comm>