Fix Supply Drop Console like OW Console for zero. (#8761)
# About the pull request
Does the same thing for Supply Drop Console as was done for OW to allow
for zero values, also brings the max and min in line with OW console
with infinities
Also allows OW Consoles to accept 0 values when saving coords.
fixes #8760
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Supply can change the height and change it back to zero.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: MistChristmas
fix: Allows Supply Drop Console to accept 0 values. Allows OW consoles
to save coords with 0 values.
/:cl: