Cleanup browser.dm (#8764)
# About the pull request

This PR is a follow up to #8734 that changed the type of a browser user
from mob to client but there were a few lingering situations where it
was not getting converted (VVing bit fields and I guess viewing
tickets).

Past that just cleanup generally in this file...

# Explain why it's good for the game

Working browsers that open?

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/be43bd03-b8ae-4a9e-afe4-13278188f7a9)

</details>


# Changelog
:cl: Drathek
fix: Fixed browsers sometimes not handling a mob passed instead of a
client (VV bitfields & tickets atleast)
/:cl: