Ports PVE Wall Leaning (#8579)
# About the pull request

Ports PVE Wall leaning to CM SS13, theres probably going to be some
combat balancing implications with this but with my light testing,
doesnt appear too bad.
 
Code by AndroBetel

# Explain why it's good for the game

Being able to lean against a wall and be all cool and mysterious is a
must have for those that actually commit to roleplay.
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/ewRo0yRVUz8
Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: AndroBetel
add: You can now lean on certain walls, click drag your sprite to the
wall and you just might. To stop a lean, either move, or press resist.
/:cl: