Ports PVE IFF Examine system (#8713)
# About the pull request

Title

# Explain why it's good for the game

Examining someone with an IFF weapon now tells you if you can shoot
through them or not.
Very handful for those without an MD and wish to know if they can fire
through a non-iff'd greeno or PMC.

# Testing Photographs and Procedure
It runs and does as intended.

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: You can now examine someone with your IFF weapon to check if the
bullet passes through or not.
/:cl: