Adds Clear Cache and Cryo and Ghost to Lobby Admin Autoreplies (#8740)
# About the pull request

As title says. Cryo and Ghost is to let people know that they need to
cryo and ghost if they wish to be lobbied.

# Explain why it's good for the game

Makes stuff easier for us. Especially clear cache.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: XQ
admin: Adds Clear Cache and Cryo and Ghost to Lobby Admin Autoreplies
/:cl: