Buffs (and renames) the M81 Riot Launcher (#8375)
# About the pull request

Replaces all instances of the M81 Riot Launcher with a new M84 Riot
Launcher, which has a non-lethal grenade capacity of 3, and the ability
to lob.

# Explain why it's good for the game

The M81 Riot Launcher always was neglected by MPs during riots. This was
because of two primary reasons
1. You cannot lob non-lethals with it like you can your own hand, to
target a back line or middle of a group.
2. You are effectively sacrificing your armor slot for what seemed like
2 extra tiles of range, and objectively more time to throw a follow up
if you needed it. This solves both of those issues, and should finally
make the new M84 launcher more relevant in riots, mutinies, and
prisonbreaks instead of your normal flashbang spam toss.

This should, as the end goal not be any more powerful than just tossing
non-lethals with your hands, but exists as an option if you need a good
and reliable way to get non-lethals where you want them and quick
without having to worry about obsessively long reload times.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/ab2d7bdc-af7f-4d8b-9452-ea7d528b3754

</details>


# Changelog
:cl:
balance: Replaced all instances of the M81 Riot Launcher with a new M84
Riot Launcher, which has a non-lethal grenade capacity of 3, and the
ability to lob.
/:cl: