Fixes bread's description (#8791)
# About the pull request
I noticed bread's description was defaulting to the base snack
description and upon further inspection realized what was probably meant
to be the description was being used as an icon_state. This is a single
word change that fixes this so the accurate description is displayed.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

![Screenshot 2025-03-17
215540](https://github.com/user-attachments/assets/400f1fa2-046d-4173-8ab3-230fe4f41d36)
![Screenshot 2025-03-17
215856](https://github.com/user-attachments/assets/51d3be1f-5c68-424e-b67e-314aa352dc00)

</details>

# Changelog

:cl:
fix: Bread now has the correct description
/:cl: