reduces the sound of queen announce (#8785)
# About the pull request
normalizes queen announce audio to a peak of -8db 

# Explain why it's good for the game

I dunno about everyone else, I have pretty sensitive ears, I run my
computer at 50 volume, and I have SS13 reduced to 50 in volume mixer on
top of that, and the sound of the xeno announce still blows the fuck out
of my ears, it's borderline painful. I pulled the file up in audacity
and the screech sounds get up to like -2db which is pretty friggin loud,
and afaik the standard for the codebase is to have sounds normalized to
-8, so I reduced peak loudness to -8 on the sound.

Obviously announcements are supposed to be attention grabbing but I
don't think they should be actively, deeply unpleasant to the point I
feel I need to mute the game while playing xeno.

<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/120d5a2c-3ed2-413d-a37d-4467d6299752

</details>


# Changelog

:cl: detectivegoogle
soundadd: queen announcement volume reduced a bit.
/:cl: