Increases Mortar Belt and Mortar Backpack Storage (#8686)
# About the pull request

Increases the number of shells able to be carried in both the mortar
belt and backpack.
Mortar belt storage: 4 > 8
Mortar backpack storage: 8 >12
While my initial focus was on the belt, it didn't make sense for the
belt and backpack to be on the same level.
# Explain why it's good for the game
These storages were not altered alongside the mortar rework and were
already useless beforehand. Since the mortar is now mobile, these
storage items should be viable.

# Testing Photographs and Procedure
I fit more shells in them
# Changelog
:cl:
balance: Increased shell storage capacity of the mortar belt and mortar
shell backpack
/:cl: