Hellhound emotes (and bug fixes) (#8574)
# About the pull request

This PR reinstates the emotes that were previously used by the Hellhound
that ended up unavailable at a certain point (it appears adding the
hellhound to the emote blacklist of the parent datum ended up completely
disabling the hellhound emotes even if there was a whitelist for the
hellhound to be able to use it). As a solution, I've simply removed the
blacklist from the parent and added the blacklist to the individual xeno
emotes that have a hellhound replacement and that do not fit with the
hellhound. I have also removed the hellhound emotes that had no sound
and were redundant due to Hellhounds being able to custom emote anyway
and produce the same result and added sounds to hiss and growl using the
LV giant lizard sounds (the hellhound already uses the giant lizard
growl when it connects a pounce so the sounds already fit)

The fixes included to the PR are:
"Find Owner" ability is now keybinded to XENO_PRIMARY_ACTION 2 so you
can use it without needing to set up a byond macro or clicking the icon
itself.

----The Hellhound no longer uses the xeno death message and xeno death
sound, now using its own death sound (same sound as its roar) and
message as was intended.---- no longer in PR (this is not the only issue
in that forsaken (not hive) xeno death code)

# Explain why it's good for the game

Bug fixes are good for the game (also more SOUL added to the game)


# Testing Photographs and Procedure
Made sure the hellhound has the emotes listed in *help and can be used,
made sure the emotes give the new sounds when activated, made sure the
death message and sounds are correct when killing the hellhound. Also
made sure I didn't mess up the death messages, emotes and sounds for the
xenos.

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:Unknownity
fix: Hellhound's "Find Owner" ability now uses the same hotkey as
XENO_PRIMARY_ACTION 2.
fix: Hellhound's emotes have returned, with improvements!
/:cl:

---------

Co-authored-by: Unknownity <a>