Increases nuke timer to 3 minutes and makes it uncancelable, nuke autotriggers after decryption is finished. (#8263)
# About the pull request
Raises the timer of **techwebs** nuke to 3 minutes and makes it
uncancelable.

Nuke autotriggers after decryption is finished.

Every time decryption is halted you lose 2 minutes of progress (to
discourage people from holding "almost decrypted" nuke).

During decryption phase Queen can reset decryption by interacting with
the nuke (just more consistent, it's weird that you would have to wait
until it's decrypted otherwise).

# Explain why it's good for the game
Nuke is meant to end the round. But command often keeps it for later
instead, still trying to win the game with regular methods while having
a reliable backup plan (the nuke). This PR aims to discourage holding
nuke for later, and to encourage using it right away, to end the round.
This would also give xenos a better shot to contest the nuke itself.

Again, nuke is meant to be a round-ender, not some soft of a backup plan
you can save for later.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
balance: raised the timer of techwebs nuke to 3 minutes and made it
uncancelable, nuke autotriggers after decryption is finished.
/:cl: