Nerfs Soro Weather effects (#8777)
# About the pull request

Makes Soro storm and blizzard effects 40% more transparent and deletes
the most contrasting pixels. (I found whatever was the least transparent
and replaced it with nothing, then applied a 40% alpha to everything)


# Explain why it's good for the game

Less visual clutter, soro is a notorious map for this and it's nearly
unplayable when it happens.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>
before:


after:

![image](https://github.com/user-attachments/assets/f5d537fe-998f-42ab-938b-079c908807de)

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl: TotalEpicness5
qol: Soro storm and blizzard effects are now 40% more transparent and
less distracting.
/:cl: