Vanguard shield minimal damage (#8773)
# About the pull request

introduces minimal damage requirement for triggering vanguard shield
decay. The main reason is so it does not get triggered by weather events
(for that minimum of 2 or so would be enough)

# Explain why it's good for the game
Having your shield that you want to use to defend yourself from rockets
triggered by snow is quite annoying and makes the shield more
predictable.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: vanguard shield requires minimal damage 6 to trigger decay
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>