db rebalance + shotgun delay cleanup (#8762)
# About the pull request

The DB exists as this unmaintained gun that's really cool but is utterly
unfit for the balance of the game, so only really exists when people
sneak one in, only to be promptly removed, this PR seeks to bring it
inline with other shotguns so it can be used as regular map decoration,
surv weapons, etc.

Currently swaps it to combat shotty tier delay, removes the burst fire
function, and generally moves its stats in line with the M37.
the sawnoff version is now just worse-- sawn off guns don't do more
damage and should generally be a downgrade in almost every way.

While I was working in this file, the delays for shotguns were all over
the place and inconsistent with some using magic numbers in SECONDS,
some using straight defines, and some using DEFINE * MAGIC NUMBER, to
rectify this a new set of defines more applicable to shotguns has been
added for use.

The merc shotgun suffers a 2 decisecond firerate nerf here in order to
bring it in line with the other combat shotties, which will occasionally
affect normal play since the gun spawns in a couple places. the merc
shotgun maintains its burst for now but I can remove it on request cuz
its frankly pretty silly too.

This PR does not change the availability of these weapons at all.
# Explain why it's good for the game


https://github.com/user-attachments/assets/e0881e52-6e87-4980-99c3-7e880cad01cb

no more of this, new colony weapon to use.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://cdn.discordapp.com/attachments/604397850675380234/1350018058336600125/Desktop_2025.03.14_-_03.06.04.08.mp4?ex=67d535cb&is=67d3e44b&hm=7c1ea3684752e6f702e9b07fa58291b36ce5443af75bedaec288e7fa5333b014&

</details>


# Changelog

:cl: detectivegoogle
balance: the alt shotguns have been rebalanced.
balance: The custom shotgun fires 2 deciseconds slower(now same as
Mk221, is otherwise the same.
balance: The double barrel(currently unobtainable) is now in most ways
equivalent to a M37, it fires as fast as a Mk221, though.
balance: The sawn-off double barrel(currently unobtainable-mostly) is
now in most ways worse than the normal DB.
code: Shotgun delay defines have been changed to be consistent, use
SECONDS for readability, and reduce magic numbers.
/:cl: