Yautja Ship - Complete Remake & more... (#8378)
# About the pull request

Completely remakes the current Hunter (Yautja) Ship, expands the size
greatly, adds multiple decks with many unique props, sprites and other
unique things.

Also adds smaller features like:

- The ability to remove the "Skull" and "Pelt" from every Xenomorph
caste, when you "Butcher"; Now it will drop both a "Skull" and "Pelt",
they will be named appropriately, giving you a much needed Trophy to
gloat over.

- Adds a unique "Trophy Rack", which let's you place all of your
Trophies on (using Pixel-shifting), works exactly like a table, and you
can even climb the lower level to reach the higher point to place your
Trophies.

Tested everything, should be fine, but it's lacking the critical
components from the "Youngblood PR", which is needs to be complete, as
it has the lower deck specifically for them and the future expansion of
the Hellhounds.

Also there's a Hangar with room for a Alamo sized dropship, it's also
something which will be expanded upon in the future possibly, but can
currently be used for events or extra space.

# Explain why it's good for the game

Much more immersive and visually interesting then the current ancient
ship we have. Many more unique sprites and things to truly make
roleplaying as a Yautja more interesting. The other decks and more space
means more expansion in the future as well, for other such Predator
related projects.

Just purely visually it's a much bigger improvement, many more sprites,
details and interesting things to look at. The main goal was just
improving the visuals, so I've removed pretty much "All" the
inappropriate sprites, basically everything that shouldn't be on a
Yautja ship, but not everything, as that would require a hell of a lot
of spriting work, but there's always room for more sprite changes in the
future and expansion.

Things I want for the future:

- Basically replacement sprites for all of the human stuff that hasn't
been touched.

- A proper Yautja Sarcophagus, which works like a morgue tray perhaps,
or just some sort of coffin.

- Replacement sprites for the ugly legacy yellow "Gladiator" armor,
something more appropriate.

- Proper Yautja Engine sprite perhaps.

- Hellhound mobs which are like the "Lizards", they should reside in the
"Hound cells".

- Any other Yautja fluff for the ship

- Proper Yautja Temple tile-set, but this is kind of unrelated.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![Hunter_Ship_WIP](https://github.com/user-attachments/assets/ab897efc-151a-41f6-8b54-654df49bcf28)


</details>


# Changelog
:cl: Zenith, Crowford, cuberound, Thwomper
code: expanded the butchering code a bit to include new items.
imageadd: added many new additions to the Hunter sprite lists and
expanded current ones, as well as many new prey items.
maptweak: completely remade the Hunter ship map and expanded it with
more decks.
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: forest2001 <forest2001@cm-ss13.com>