Removes RMC rifle from hunting grounds, replaces it with AUG (#8627)
# About the pull request

Replaces the RMC L23 with an Aug on hunting grounds

# Explain why it's good for the game

Originally i was going to nerf the RMC rifle but then i decided that
given almost all the complaints are about it in hunting grounds and that
if i were to rebalance one ERT gun i should do them all at the same
time, i decided to just replace it with the original AUG.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:

balance: Lowered the L23's damage and gave it falloff.
/:cl: