Tgui update 1 (#8676)
# About the pull request


@hry-gh This should be tested first, best test tgui say and some uis

webpack to edge only, requires 516

ports https://github.com/tgstation/tgstation/pull/88919 adjusted for
cmss, please check that the functionality is all there.

# Explain why it's good for the game
prepares tgui for the tgui-core migration, I'll keep all the components
and themes imported from the local repo, it's up to you to swap or not
in the end.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
code: update tgui dependencies
code: update tgui say
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>