Makes the green beret in SO vendors actually green (#8907)
# About the pull request
Title sums it up. Makes the "Beret, Green" option in the SO vendor give
a `beret/cm/green` instead of `beret/cm`
# Explain why it's good for the game
The vendor says a green beret but spits out a beret that changes color.
This makes it a green beret.
# Changelog
:cl:
fix: The Staff Officer vendor's green beret is now actually green.
/:cl: