Universal speak works now. (#8868)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
makes the code better, there was a lot of dead code, and fixes universal
speak.
also moves a lot of 1 to true,etc.
Also removed universal speak being also universal understand, because
its not very intuitive.
# Explain why it's good for the game
I switched universal speak to true in tdome, and it didnt work. now Im
sad so I fix it.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: universal speak is working again
/:cl: