Fixes CO pistols being melted or exploded (#8866)
# About the pull request

Title

# Explain why it's good for the game

COs really shouldn't be losing these things. Just because it's unlikely
that a BE would happen doesn't mean you should be locked out of it for
the round.

Just so we are clear, the Mateba has always never been able to be
melted, this just standardizes the logic.


# Testing Photographs and Procedure

Line change, should work

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: CO Pistols no longer get melted or blown up.
/:cl: