[Code bounty] Yautja Hunting Bow 2 (#8763)
# About the pull request
Previous PR died to merge conflicts.

**Design doc**:
https://docs.google.com/document/d/1MaMh5aBPh8d8v-qnHQBZf3iv9qrfyg8RnjJ8FNVJaPs/edit?tab=t.0

Adds a bow, a quiver and arrows, and icons for everything. 

Only yautja can use the bow. The weapon is designed to be used against
dishonorable prey. Arrows have two modes: standard and activated. The
standard arrow acts like an arrow, it simply deals damage on impact and
gets stuck in the body (you can pull it out of the body later, if it's a
human). If you miss, the arrow remains on the ground. The activated
arrow explodes on impact, which is similar to lethal HPC, cannot be
recovered. The arrow can fall out of the bow if you drop the bow or try
to return the bow to the strap, which means you can not store a loaded
bow on the strap. The quiver strap holds a bow and 7 arrows, it can be
worn on the belt and in the suit storage slots.

I also changed "yank out" verb slightly for it to make a bit more sense.

All credits to @thwompa 

# Explain why it's good for the game
More options for yautja. Cool sprites. Interesting mechanics.

# Testing Photographs and Procedure
Videos are a bit outdated, but should give you the general idea.


https://github.com/user-attachments/assets/d3a60e29-6f9b-4c1a-9fcf-f14810578dfb


https://github.com/user-attachments/assets/d827d7a6-b5c8-462b-9bcc-eede0391659d

# Changelog
:cl: Thwomper, ihatethisengine, DawsonKeyes
add: introduced hunting bow as a secondary weapon for yautja.
balance: yanking out objects speed now depends on the size of the object
/:cl: