Gun durability and jamming system (#8682)
# About the pull request

As per the title.
Adds a durability and jamming system to weapons and the means to
maintain it.

# Explain why it's good for the game

In order for marines to proceed to more emergent gameplay for roadmap
2025, there should be more measures to prevent xenos from melting your
gun immediately and then being out of combat for upwards to 30 minutes,
(Exaggeration) depending on the competence of other players.

Frankly, there should be less qdeletion of weapons, or items in general,
and a more comprehensive repair system, but this is currently what is
cooked up.

Code is based on Carlarcs PPSH and UZI jamming code that doesn't even
work.

Essentially Feature complete.
Next PRs will probably include:
Misfiring.
Fire damaging durability
Refactored acid code

# Testing Photographs and Procedure
<details>

<summary>Screenshots & Videos</summary>


https://cdn.discordapp.com/attachments/604397850675380234/1345337020699902002/Desktop_2025.03.01_-_18.08.04.03.mp4?ex=67cac5bc&is=67c9743c&hm=eb1b2b86a0bf5b21399d98cf47b6322a50f6862fe6f9e604ed1853dfc1ee4338&

</details>


# Changelog
:cl:
add: Adds a gun durability and repair system.
add: A handful of bullets can be used to chamber a gun when it is
unchambered, provided the caliber matches.
balance: Guns can no longer be destroyed by small explosions; this
functionality is hooked up with the gun durability system, but guns can
still be deleted if the explosion is strong enough i.e. Orbital
Bombardments.
balance: Guns can no longer be deleted when acid is poured into it,
instead requiring repairs. Zero durability weapons still get deleted if
its acided, however.
/:cl:

---------

Co-authored-by: Detective-Google <48196179+Detective-Google@users.noreply.github.com>