Adds Smooth marine armor to Surplus Uniform Vendor + Grammar Tweak (#8956)
# About the pull request
I noticed every other type of medium armor was in the vendor except
Smooth so I added it in, and also changed the M10 extended mag and drum
mag names to match the capitalization of the rest of the vendor.
# Explain why it's good for the game
It seems odd that one of the purely visual variants of armor is missing
from the vendor.
# Changelog

:cl:
add: Added M3 Smooth Marine Armor to Req's Surplus Uniform Vendor
spellcheck: Tweaked the capitalization of the M10 magazines in the Req
Ammo Vendor.
/:cl: