Fixes a small error in bone-mending description (#8965)
# About the pull request

Fixes a 1 character spelling mistake in the bone-mending chem property,
"Re-wowen" to "Re-woven"

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/8f0048fd-c3b3-449e-bee6-2a51f4c506d0)

</details>


# Changelog
:cl: Doc12693420
spellcheck: fixes a typo in bone-mending chem property, "Re-wowen" to
"Re-woven"
/:cl: