SSV Rostock - UPP Warship Map (#7525)
# About the pull request

This PR adds the SSV Rostock, a UPP warship, as a template that can be
spawned in by admins.

This is a fully mapped warship intended for use by the UPP, it is
intended to allow an entire team of UPP to be spawned in, gear up, and
deploy to a combat map/arena via dropships. It contains similar onboard
amenities to the USS Almayer, including a fully stocked medbay, cargo
bay, MP department and CIC.

It also features a functional ASRS system.

Other changes

- Adds features of this PR:
https://github.com/cmss13-devs/cmss13/pull/7548
- Adds heavy cable decal sprites for mapper use.
- Adds UPP wall/floor sprites (credit to Zenith for the retextured
sprites, and dimdimich1996 for the custom ship sprites.)
- Additional code changes to better facilitate integration of HvH
gamemode features, such as ceasefire times, game end conditions,
automatic loading of the Rostock map.
- New modifier that prevents all humans from using weapons. Used
automatically in the HvH gamemode, can be used by admins.
- Code to allow fusion reactors to not start broken when spawned on
non-ground maps.
- UPP medical vendors. 
- UPP lifeboat

# Explain why it's good for the game

See attached document for custom-made lore documentation in regards to
this specific vessel, the Rostock, but also its class of ship in
general, the Brandenburg-Class Expeditionary Frigate: [Rostock
Lore.txt](https://github.com/user-attachments/files/17778101/Rostock.Lore.txt)


This ship will significantly increase the ability for server staff to
run events for the UPP. Either to host a HvH event, to spawn the UPP as
a alternative to the USCM for distress call, or to use the map as a
small combat arena for the Xenomorphs or any other third party to use,
either as a side mission or a thunder dome alternative.

The vessel does require some admin-intervention to make it fully
operational, such as.

1. A lifeboat spawns, but the lifeboat can not be activated or launched
without admin intervention.

But in all other capacities the ship is fully functional, if you were to
drop a team of OpFor with no gear, they would be able to be fully geared
up.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>
Very slightly out of date image


![uppship_02_11](https://github.com/user-attachments/assets/61984106-c98f-435a-8123-a57e99d9433b)


</details>


# Changelog
:cl: Steelpoint, Cuberound, Blundir, Zenith, dimdimich1996, .gl0ck,
feroxz
mapadd: The SSV Rostock, a UPP warship equivalent to the USS Almayer,
has been added to the game as an admin-spawnable ship.
mapadd: A UPP lifeboat has been added. 
add: The SSV Rostock will automatically be spawned in during the
'Faction Clash UPP vs USCM' gamemode, as the primary base of operations
for the UPP. This vessel will have spawners allowing the UPP to spawn in
from the lobby on to the Rostock.
admin: In order for the ship to be loaded, a new Verb in the Server
panel has been added. The M: Prepare Events verb. This will restart the
server once a gamemode has been selected.
add: The Rostock will also have a set of UPP dropships, and a custom
lifeboat spawned on the vessel when it is loaded for the Faction Clash
game mode.
add: UPP Major and Minor victories have been added. 
admin: If the 'end-round' function is used, any one of all the victory
condition can now be set, instead of the game automatically being
considered a draw.
admin: A new modifier has been added, the 'ceasefire'. Prevents all
humans from using any firearm or explosives. This is automatically used
during the 'Faction Clash' gamemode, and can be activated manually be
server staff for events.
code: Conditions for a ceasefire period have been added for the 'Faction
Clash UPP USCM' game mode. This institutes a five-minute ceasefire
preventing both factions from attacking the other once any faction
transport lands on the operations zone. The commencement and cessation
of this ceasefire are automatically announced.
code: Once all forces of a single faction have been eliminated on the
operations zone during the 'Faction Clash UPP USCM' game mode, the game
will automatically declare the other faction as the victory, which will
end the round.
add: UPP medical vendors have been added to the game. 
add: A suite of UPP ship wall and hull tiles have been added to the game
for mapper usage.
add: Heavy cable sprites have been added as a decal for better mapper
use.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>
Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>
Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: cuberound <122645057+cuberound@users.noreply.github.com>
Co-authored-by: gl0ck <legobro219@gmail.com>
Co-authored-by: Aurallianz <Feroxtheheist@protonmail.com>